+++ shim.js
@@ -1,0 +1,2 @@
+/* used when calling require('htmlvalidate'); */
+module.exports = require('./build/htmlvalidate').default;