+++ src/transform/passthru.ts
@@ -1,0 +1,5 @@
+const fs = require('fs');
+
+module.exports = function passthruTransform(filename: string) {
+	return fs.readFileSync(filename, {encoding: 'utf8'});
+};